backend/pyproject.toml
[project]
name = "rag-chatbot-backend"
version = "0.1.0"
description = "Backend for RAG Chatbot with Cohere and Qdrant integration"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "cohere>=4.4.3",
    "qdrant-client>=1.8.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]